Smart Document Assistant is Public

+++ src/main/java/gae/piaz/sda/repository/DocumentRepository.java
@@ -4,4 +4,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DocumentRepository extends JpaRepository<DocumentEntity, Long> { }
+public interface DocumentRepository extends JpaRepository<DocumentEntity, Integer> {}
